# excel-api/api-docs.yaml
openapi: 3.0.3
info:
  title: LH Calculator API
  version: 1.0.0
  description: Excel-based calculation service with 134+ fields

servers:
  - url: http://localhost:3000
    description: Local development

paths:
  /api/calculate:
    post:
      summary: Process calculation
      operationId: calculate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculationRequest'
      responses:
        '200':
          description: Successful calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationResponse'
        '422':
          description: Validation or processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable (queue full)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"

  /api/metrics:
    get:
      summary: System metrics
      responses:
        '200':
          description: Current system metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  memory:
                    type: object
                  queue:
                    type: object
                  workers:
                    type: object

components:
  schemas:
    CalculationRequest:
      type: object
      required:
        - tech_D27_sequenceNumber
        - tech_E27_customerOrderPosition
        - tech_H27_passes
        - tech_I27_plateQuantity
        - tech_J27_calcPressureHotSide
        - tech_K27_calcPressureColdSide
        - tech_L27_calcTempHotSide
        - tech_M27_calcTempColdSide
        - tech_T27_drawDepth
        - sup_F2_projectNumber
        - sup_D8_flowPartMaterialPricePerKg
        - sup_E8_flowPartMaterialPrice
        - sup_K13_normHoursPerUnit
        - sup_P13_internalLogistics
        - sup_D78_stainlessSteelThickness
      properties:
        tech_D27_sequenceNumber:
          type: number
          minimum: 0
        tech_E27_customerOrderPosition:
          type: string
        tech_H27_passes:
          type: string
        tech_I27_plateQuantity:
          type: number
          minimum: 0
        tech_J27_calcPressureHotSide:
          type: number
          minimum: 0
        tech_K27_calcPressureColdSide:
          type: number
          minimum: 0
        tech_L27_calcTempHotSide:
          type: number
          minimum: 0
        tech_M27_calcTempColdSide:
          type: number
          minimum: 0
        tech_T27_drawDepth:
          type: number
          minimum: 0
        sup_F2_projectNumber:
          type: string
          enum: ['0000', '06ХН28МДТ', '09Г2С', '12Х18Н10Т', '20ХН3А', '30ХМА', '40Х']
        sup_D8_flowPartMaterialPricePerKg:
          type: number
          minimum: 0
        sup_E8_flowPartMaterialPrice:
          type: number
          minimum: 0
        sup_K13_normHoursPerUnit:
          type: number
          minimum: 0
        sup_P13_internalLogistics:
          type: number
          minimum: 0
        sup_D78_stainlessSteelThickness:
          type: number
          minimum: 0
      additionalProperties: true

    CalculationResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        results:
          type: object
          properties:
            calculated_values:
              type: object
            total_cost:
              type: number
            component_costs:
              type: object
              properties:
                materials:
                  type: number
                processing:
                  type: number
                hardware:
                  type: number
                other:
                  type: number
        request_id:
          type: string
        processing_time_ms:
          type: number

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        details:
          type: object
        request_id:
          type: string